% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPI.R
\name{GPI}
\alias{GPI}
\title{Generalized Probability Inference (GPI)}
\usage{
GPI(
  data,
  control_group = NULL,
  dir_mul = TRUE,
  dispersion = list(psi_mean = 0, psi_sd = 0.5, varphi_mean = 0, varphi_sd = 0.01),
  control = list(chains = 3, iter_count = 1000, iter_V = 10000, cores = 3, warm = FALSE),
  q_up = 0.975,
  q_lo = 0.025,
  n_sam = NULL,
  in_vivo = TRUE,
  ...
)
}
\arguments{
\item{data}{A list containing the following components:
\itemize{
  \item \code{data_count}: A matrix or data frame of barcode read counts. Rows are experimental replicates (e.g., DMSO_1), and columns are genotypes or mutations (e.g., PTEN, TSC2).
  \item \code{condition_count}: A factor vector specifying treatment groups for \code{data_count}.
  \item \code{V}: A numeric vector of either tumor volumes (in vivo) or confluency values (in vitro).
  \item \code{condition_v}: A factor vector specifying treatment groups for \code{V}.
  \item \code{cell_line}: A factor vector of cell lines corresponding to the replicates in \code{data_count}.
  \item \code{VT0}: Optional numeric vector of baseline tumor volumes or confluency values for normalization.
}}

\item{control_group}{Control group for comparative analysis.}

\item{dir_mul}{Logical flag indicating the choice of model for the Dirichlet-multinomial likelihood:
\itemize{
  \item \code{TRUE}: Use the model defined in \code{dir_mul.stan}.
  \item \code{FALSE}: Use the model defined in \code{dir_new.stan}.
}}

\item{dispersion}{A list of parameters to control dispersion in the Dirichlet-multinomial model:
\itemize{
  \item \code{psi_mean}: Mean for the log-normal prior on the precision parameter \code{psi} (default: 0).
  \item \code{psi_sd}: Standard deviation for the log-normal prior on \code{psi} (default: 0.5).
  \item \code{varphi_mean}: Mean for the log-normal prior on the variance parameter \code{varphi} (default: 0).
  \item \code{varphi_sd}: Standard deviation for the log-normal prior on \code{varphi} (default: 0.01).
}}

\item{control}{List of control parameters:
\itemize{
  \item \code{chains}: How many Markov chains should be used for model fitting (default: 3).
  \item \code{iter_count}: #'     \item \code{iter_count}: Number of iterations for fitting the barcoding data. Increasing this value may improve the effective sample size (default: 1000).
  \item \code{iter_V}: Number of iterations for fitting the tumor volume or confluency data. Increasing this value may improve the effective sample size (default: 10000).
  \item \code{cores}: Number of CPU cores to use (default: 3).
  \item \code{warm}: Logical, indicating whether warm-up iterations should be included (default: FALSE).
}}

\item{q_up}{Upper quantile threshold (default: 0.975).}

\item{q_lo}{Lower quantile threshold (default: 0.025).}

\item{n_sam}{Number of posterior samples (defaults to \code{iter / 2}).}

\item{in_vivo}{Logical flag indicating whether the data are from an in vivo experiment (default: \code{TRUE}).
If \code{TRUE}, a log-normal likelihood is used for tumor volume. If \code{FALSE}, a beta likelihood is used for confluency.}

\item{...}{Other parameters for model fitting.}
}
\value{
A list containing fitted Stan models and relevant metadata:
  \itemize{
    \item \code{fit_dir_mult}: Fitted Dirichlet-multinomial model.
    \item \code{fit_V}: Fitted tumor volume or confluency model.
    \item \code{L}: Number of unique cell lines.
    \item \code{iter_count}, \code{iter_V}, \code{chains}: MCMC parameters used.
    \item \code{data_count}, \code{V}: Input data.
    \item \code{group}, \code{group_V}: Numeric group identifiers.
    \item \code{condition_count}, \code{condition_v}: Group conditions for count and volume data.
    \item \code{control_group}: Control group identifier.
    \item \code{n_sam}: Number of posterior samples.
  }
}
\description{
Conducts Bayesian inference on barcoding data and tumor volume (in vivo) or confluency (in vitro) measurements.
Utilizes a Dirichlet-multinomial model for count data, a log-normal likelihood for tumor volume in vivo,
and a beta likelihood for confluency in vitro.
}
